SQ80：change KB_folder；fix some code style questions
change the name of  keyboard folder
using "#pragma once" instead of "#if"
using functions for GPIO form quantum.h  instead the AVRstyle codes
using “layout” instead of “keymap”
add readme.mk
add come information  about BL to the  rule.mk
