+++ keyboards/inett_studio/sq80/config.h
@@ -1,5 +1,3 @@
-#ifndef CONFIG_H
-#define CONFIG_H
 #pragma once
 #include "config_common.h"
 
@@ -30,6 +28,4 @@
 #define BACKLIGHT_PIN B7
 #ifdef BACKLIGHT_PIN
 #define BACKLIGHT_LEVELS 5
-#endif
-
 #endif
