+++ keyboards/inett_studio/sq80/readme.md
@@ -1,0 +1,15 @@
+# SQ80
+
+![SQ80](imgur.com image replace me!)
+
+A short description of the keyboard/project
+
+Keyboard Maintainer: QMK Community  
+Hardware Supported: SQ80 and SQ80wkl
+Hardware Availability: iNETT STUDIO   QQ群号：528509432
+
+Make example for this keyboard (after setting up your build environment):
+
+    make pitta/sq80:default
+
+See the [build environment setup](https://docs.qmk.fm/#/getting_started_build_tools) and the [make instructions](https://docs.qmk.fm/#/getting_started_make_guide) for more information. Brand new to QMK? Start with our [Complete Newbs Guide](https://docs.qmk.fm/#/newbs).
