+++ keyboards/inett_studio/sq80/sq80.h
@@ -1,9 +1,7 @@
-#ifndef sq80_H
-#define sq80_H
-
+#pragma once
 #include "quantum.h"
 
-#define KEYMAP_ANSI( \
+#define LAYOUT_ansi( \
 	K000, K001, K002, K003, K004, K005, K006, K007, K008, K009, K010, K011, K012, K013, K014, \
 	K100, K101, K102, K103, K104, K105, K106, K107, K108, K109, K110, K111, K112, K113, K114, \
 	K200, K201, K202, K203, K204, K205, K206, K207, K208, K209, K210, K211, K212, K213, K214, \
@@ -19,8 +17,7 @@
 	{ K500,  K501,  K502,  K503,  K504,  K505,  K506,  K507,  K508,  K509,  K510,  K511,  K512,  K513,  K514 }  \
 }
 
-
-#define KEYMAP_WKL( \
+#define LAYOUT_wkl( \
 	K000, K001, K002, K003, K004, K005, K006, K007, K008, K009, K010, K011, K012, K013, K014, \
 	K100, K101, K102, K103, K104, K105, K106, K107, K108, K109, K110, K111, K112, K113, K114, \
 	K200, K201, K202, K203, K204, K205, K206, K207, K208, K209, K210, K211, K212, K213, K214, \
@@ -35,6 +32,3 @@
 	{ K400,  K401,  K402,  K403,  K404,  K405,  K406,  K407,  K408,  K409,  K410,  KC_NO, KC_NO, K413,  K414 }, \
 	{ K500,  K501,  K502,  K503,  K504,  K505,  K506,  K507,  KC_NO, K509,  K510,  K511,  K512,  K513,  K514 }  \
 }
-
-
-#endif
